convert sqlte to mysql (doing)

+++ app/src/main/java/com/example/foodapplication/Transaction.java
@@ -3,14 +3,14 @@ package com.example.foodapplication;
 import java.util.Date;
 
 public class Transaction {
-    private int credits;
+    private double credits;
     private String date;
 
-    public Transaction(int credits, String date) {
+    public Transaction(double credits, String date) {
         this.credits = credits;
         this.date = date;
     }
 
-    public int getCredits() { return credits;}
+    public double getCredits() { return credits;}
     public String getDate() { return date; }
 }
